access rights image done

+++ app/MagicMonkey/MiniJournal/Controller/ImageController.php
@@ -30,10 +30,18 @@ class ImageController extends AbstractController
      * 1) récupère toutes les images
      * 2) affichage
      */
-    public function listAll()
+    public function publiees()
     {
-        $lstObjsImages = (new ImageBd())->selectAll();
-        $this->render("image/vAllImages.html.twig", array("images" => $lstObjsImages));
+        $lstImages = (new ImageBd())->selectAllBy(array('publication_status =' => 'publie'));
+        $this->render("image/vAllImages.html.twig", array("endH1" => "publiées", "images" => $lstImages));
+    }
+
+    public function nonPubliees()
+    {
+        if ($this->roleManager->renderAccessDenied($this, array("ROLE_ADMIN", "ROLE_EDITOR"))) {
+            $lstImages = (new ImageBd())->selectAllBy(array('publication_status =' => 'brouillon'));
+            $this->render("image/vAllImages.html.twig", array("endH1" => "non publiées", "images" => $lstImages));
+        }
     }
 
     /**
@@ -41,37 +49,40 @@ class ImageController extends AbstractController
      */
     public function insert()
     {
-        $imageForm = new ImageForm();
-        $imageeBd = new ImageBd();
-        if (count($this->request->getPost()) == 0) { // s'il n'y a pas de données postées
-            $this->render("image/vFormNewUpdate.html.twig", array( // affichage du formulaire d'ajout d'une image
-                "title" => "Ajout d'une image",
-                "action" => "insert",
-                "imageForm" => $imageForm
-            ));
-        } else {
-            // récupération et manipulation du file image
-            $file = $this->request->getFilesParam("image");
-            $fm = new FileManager();
-            $infoImg = $file['tmp_name'] ? getimagesize($file['tmp_name']) : false;
-            $this->request->addPostParam("file", array($file, $infoImg));
-            if ($imageForm->validate($this->request->getPost())) { // verif du formulaire : si aucune erreur
-                $this->request->removePostParam('file');
-                /* upload grâce à l'objet fm FileManager */
-                $this->request->addPostParam("path", $fm->upload($file, DIR_IMAGE));
-                $imageeBd->add($this->request->getPost()); /* enregistrement du nouvel article dans la bdd */
-                $_SESSION['success'] = "Ajout effectué !";
-                header('Location: index.php?a=listAll&o=image'); // redirection vers la liste des images
-                exit();
-            } else { // s'il y a au moins une erreur
-                $imageForm->setImage($imageeBd->mapp($this->request->getPost()));
-                // redirection vers le formulaire avec notification
-                $this->render("image/vFormNewUpdate.html.twig", array(
-                    "imageForm" => $imageForm,
+        if ($this->roleManager->renderAccessDenied($this, array("ROLE_EDITOR", "ROLE_ADMIN"))) {
+            $imageForm = new ImageForm();
+            $imageeBd = new ImageBd();
+            if (count($this->request->getPost()) == 0) { // s'il n'y a pas de données postées
+                $this->render("image/vFormNewUpdate.html.twig", array( // affichage du formulaire d'ajout d'une image
                     "title" => "Ajout d'une image",
-                    "image" => $imageForm->getImage(),
-                    "action" => "insert"
+                    "action" => "insert",
+                    "imageForm" => $imageForm
                 ));
+            } else {
+                // récupération et manipulation du file image
+                $file = $this->request->getFilesParam("image");
+                $fm = new FileManager();
+                $infoImg = $file['tmp_name'] ? getimagesize($file['tmp_name']) : false;
+                $this->request->addPostParam("file", array($file, $infoImg));
+                if ($imageForm->validate($this->request->getPost())) {
+                    // verif du formulaire : si aucune erreur
+                    $this->request->removePostParam('file');
+                    /* upload grâce à l'objet fm FileManager */
+                    $this->request->addPostParam("path", $fm->upload($file, DIR_IMAGE));
+                    $imageeBd->add($this->request->getPost()); /* enregistrement du nouvel article dans la bdd */
+                    $_SESSION['success'] = "Ajout effectué !";
+                    header('Location: index.php?a=publiees&o=image'); // redirection vers la liste des images
+                    exit();
+                } else { // s'il y a au moins une erreur
+                    $imageForm->setImage($imageeBd->mapp($this->request->getPost()));
+                    // redirection vers le formulaire avec notification
+                    $this->render("image/vFormNewUpdate.html.twig", array(
+                        "imageForm" => $imageForm,
+                        "title" => "Ajout d'une image",
+                        "image" => $imageForm->getImage(),
+                        "action" => "insert"
+                    ));
+                }
             }
         }
     }
@@ -81,42 +92,51 @@ class ImageController extends AbstractController
      */
     public function delete()
     {
-        $imageBd = new ImageBd();
-        $imageForm = new ImageForm();
-        /* stockage de tests dans des variables */
-        $isEmptyGetId = !isset($this->request->getGet()['id']);
-        $isNotEmptyPostImage = isset($this->request->getPost()['image']);
-        /* utilisation des tests */
-        if (($isEmptyGetId && $isNotEmptyPostImage) || !$isEmptyGetId) {
-            $id = $isEmptyGetId ? $this->request->getPost()['image'] : $this->request->getGet()['id'];
-            $imageForm->setImage($imageBd->selectOne(array("id =" => (int)$id)));
-            $res = $imageBd->deleteOne($id);
-            if (!$res && $isNotEmptyPostImage) { // identifiant non renseigné
-                $imageForm->setErrors(array("errorImage" => "L'image doit être indiquée et doit existée"));
+        if ($this->roleManager->renderAccessDenied($this, array('ROLE_ADMIN', 'ROLE_EDITOR'))) {
+            $imageBd = new ImageBd();
+            $imageForm = new ImageForm();
+            /* stockage de tests dans des variables */
+            $isEmptyGetId = !isset($this->request->getGet()['id']);
+            $isNotEmptyPostImage = isset($this->request->getPost()['image']);
+            /* utilisation des tests */
+            if (($isEmptyGetId && $isNotEmptyPostImage) || !$isEmptyGetId) {
+                $id = $isEmptyGetId ? $this->request->getPost()['image'] : $this->request->getGet()['id'];
+                $image = $imageBd->selectOne(array('id =' => (int)$id));
+                if ($image && ($image->getPublicationStatus() == "brouillon") || $this->roleManager->isAuth()) {
+                    if ($this->roleManager->renderAuthorDenied($this, $image)) {
+                        $imageForm->setImage($imageBd->selectOne(array("id =" => (int)$id)));
+                        $res = $imageBd->deleteOne($id);
+                        if (!$res && $isNotEmptyPostImage) { // identifiant non renseigné
+                            $imageForm->setErrors(array("errorImage" => "L'image doit être indiquée et doit existée"));
+                            $this->render("image/vFormSelect.html.twig", array(
+                                "images" => $imageBd->selectAll(),
+                                "imageForm" => $imageForm,
+                                "action" => "delete"
+                            ));
+                        } else {
+                            if (!$res && !$isEmptyGetId) { // l'image n'existe pas
+                                $_SESSION['error'] = "Une erreur est survenue lors de la suppression de
+                       l'image ! Il se peut que l'image n'existe plus";
+                            } else {
+                                // suppression du fichier => redirection vers la liste des images
+                                $fm = new FileManager();
+                                $fm->deleteFile($imageForm->getImage()->getPath(), DIR_IMAGE);
+                                $_SESSION['success'] = "Suppression effectuée !";
+                            }
+                            header('Location: index.php?a=publiees&o=image');
+                            exit();
+                        }
+                    }
+                } else {
+                    $this->render("image/vCantDeletePublished.html.twig");
+                }
+            } else { // affichage du formulaire pour choisir une image à supprimer
                 $this->render("image/vFormSelect.html.twig", array(
                     "images" => $imageBd->selectAll(),
                     "imageForm" => $imageForm,
                     "action" => "delete"
                 ));
-            } else {
-                if (!$res && !$isEmptyGetId) { // l'image n'existe pas
-                    $_SESSION['error'] = "Une erreur est survenue lors de la suppression de
-                       l'image ! Il se peut que l'image n'existe plus";
-                } else {
-                    // suppression du fichier => redirection vers la liste des images
-                    $fm = new FileManager();
-                    $fm->deleteFile($imageForm->getImage()->getPath(), DIR_IMAGE);
-                    $_SESSION['success'] = "Suppression effectuée !";
-                }
-                header('Location: index.php?a=listAll&o=image');
-                exit();
             }
-        } else { // affichage du formulaire pour choisir une image à supprimer
-            $this->render("image/vFormSelect.html.twig", array(
-                "images" => $imageBd->selectAll(),
-                "imageForm" => $imageForm,
-                "action" => "delete"
-            ));
         }
     }
 
@@ -129,14 +149,14 @@ class ImageController extends AbstractController
             $image = (new ImageBd())->selectOne(array("id =" => (int)$this->request->getGet()['id']));
             if (!$image) { /* image inexistante */
                 $_SESSION['error'] = "L'image demandée n'existe pas";
-                header('Location: index.php?a=listAll&o=image');
+                header('Location: index.php?a=publiees&o=image');
                 exit();
             } else { /* image existe */
                 $this->render("image/vOneImage.html.twig", array("image" => $image));
             }
         } else { /* identifiant non renseigné */
             $_SESSION['error'] = "Aucune image demandée";
-            header('Location: index.php?a=listAll&o=image');
+            header('Location: index.php?a=publiees&o=image');
             exit();
         }
     }
@@ -146,65 +166,69 @@ class ImageController extends AbstractController
      */
     public function update()
     {
-        $imageBd = new ImageBd();
-        $imageForm = new ImageForm();
-        if (!isset($this->request->getGet()['id'])) { // id non renseigné
-            $_SESSION['error'] = "Aucune image demandée";
-            header('Location: index.php?o=image&a=listAll');
-            exit();
-        } else {
-            $imageForm->setImage($imageBd->selectOne(array("id =" => (int)$this->request->getGet()['id']), false));
-            if (count($this->request->getPost()) == 0) { // s'il n'y a pas de données postées
-                if ($imageForm->getImage()) { // si l'image existe => affichage du formulaire de modification
-                    $this->render("image/vFormNewUpdate.html.twig", array(
-                        "title" => "Modification d'une image",
-                        "image" => $imageForm->getImage(),
-                        "action" => "update&o=image&id=" . $imageForm->getImage()->getId(),
-                        "imageForm" => $imageForm
-                    ));
-                } else { // si elle n'existe pas
-                    $_SESSION['error'] = "L'image demandée n'existe pas";
-                    // redirection => liste des images avec notifications
-                    header('Location: index.php?o=image&a=listAll');
-                    exit();
-                }
-            } else { // s'il y a des données postées
-                // manipulation du fichier image
-                $file = $this->request->getFilesParam("image");
-                $fm = new FileManager();
-                $infoImg = $file['tmp_name'] ? getimagesize($file['tmp_name']) : true;
-                $this->request->addPostParam("file", array($file, $infoImg));
-                if ($imageForm->validate($this->request->getPost())) { // verif du formulaire : si aucune erreur
-                    /* si post file infoImg == true */
-                    $test = $this->request->getPostParam('file', false);
-                    if (!is_array($test[1])) {
-                        /* pas besoin d'upload une nouvelle image */
-                        /* add post param 'path' avec l'ancienne valeur soit  $image->getPath()*/
-                        $this->request->addPostParam('path', $imageForm->getImage()->getPath());
-                    } else { /* sinon */
-                        /* on upload gràace à l'objet $fm FileManager */
-                        $this->request->addPostParam("path", $fm->upload($file, DIR_IMAGE));
-                        /* et on delete l'ancienne */
-                        $fm->deleteFile($imageForm->getImage()->getPath(), DIR_IMAGE);
+        if ($this->roleManager->renderAccessDenied($this, array('ROLE_ADMIN', 'ROLE_EDITOR'))) {
+            $imageBd = new ImageBd();
+            $imageForm = new ImageForm();
+            if (!isset($this->request->getGet()['id'])) { // id non renseigné
+                $_SESSION['error'] = "Aucune image demandée";
+                header('Location: index.php?o=image&a=publiees');
+                exit();
+            } else {
+                $imageForm->setImage($imageBd->selectOne(array("id =" => (int)$this->request->getGet()['id']), false));
+                if ($this->roleManager->renderAuthorDenied($this, $imageForm->getImage())) {
+                    if (count($this->request->getPost()) == 0) { // s'il n'y a pas de données postées
+                        if ($imageForm->getImage()) { // si l'image existe => affichage du formulaire de modification
+                            $this->render("image/vFormNewUpdate.html.twig", array(
+                                "title" => "Modification d'une image",
+                                "image" => $imageForm->getImage(),
+                                "action" => "update&o=image&id=" . $imageForm->getImage()->getId(),
+                                "imageForm" => $imageForm
+                            ));
+                        } else { // si elle n'existe pas
+                            $_SESSION['error'] = "L'image demandée n'existe pas";
+                            // redirection => liste des images avec notifications
+                            header('Location: index.php?o=image&a=publiees');
+                            exit();
+                        }
+                    } else { // s'il y a des données postées
+                        // manipulation du fichier image
+                        $file = $this->request->getFilesParam("image");
+                        $fm = new FileManager();
+                        $infoImg = $file['tmp_name'] ? getimagesize($file['tmp_name']) : true;
+                        $this->request->addPostParam("file", array($file, $infoImg));
+                        if ($imageForm->validate($this->request->getPost())) { // verif du formulaire : si aucune erreur
+                            /* si post file infoImg == true */
+                            $test = $this->request->getPostParam('file', false);
+                            if (!is_array($test[1])) {
+                                /* pas besoin d'upload une nouvelle image */
+                                /* add post param 'path' avec l'ancienne valeur soit  $image->getPath()*/
+                                $this->request->addPostParam('path', $imageForm->getImage()->getPath());
+                            } else { /* sinon */
+                                /* on upload gràace à l'objet $fm FileManager */
+                                $this->request->addPostParam("path", $fm->upload($file, DIR_IMAGE));
+                                /* et on delete l'ancienne */
+                                $fm->deleteFile($imageForm->getImage()->getPath(), DIR_IMAGE);
+                            }
+                            /* on supprime le file post */
+                            $this->request->removePostParam('file');
+                            /* modification de l'image dans la bdd */
+                            $imageBd->update($this->request->getPost(), $imageForm->getImage()->getId());
+                            $_SESSION['success'] = "Modification effectuée";
+                            header('Location: index.php?o=image&a=describe&id=' . $imageForm->getImage()->getId());
+                            exit();
+                        } else {
+                            // s'il y a au moins une erreur => redirection vers le formulaire de modification avec des
+                            // notifications pour aider l'utilisateur
+                            $this->request->addPostParam('id', $imageForm->getImage()->getId());
+                            $imageForm->setImage($imageBd->mapp($this->request->getPost()));
+                            $this->render("image/vFormNewUpdate.html.twig", array(
+                                "title" => "Modification d'une image",
+                                "image" => $imageForm->getImage(),
+                                "action" => "update&o=image&id=" . $imageForm->getImage()->getId(),
+                                "imageForm" => $imageForm
+                            ));
+                        }
                     }
-                    /* on supprime le file post */
-                    $this->request->removePostParam('file');
-                    /* modification de l'image dans la bdd */
-                    $imageBd->update($this->request->getPost(), $imageForm->getImage()->getId());
-                    $_SESSION['success'] = "Modification effectuée";
-                    header('Location: index.php?o=image&a=describe&id=' . $imageForm->getImage()->getId());
-                    exit();
-                } else {
-                    // s'il y a au moins une erreur => redirection vers le formulaire de modification avec des
-                    // notifications pour aider l'utilisateur
-                    $this->request->addPostParam('id', $imageForm->getImage()->getId());
-                    $imageForm->setImage($imageBd->mapp($this->request->getPost()));
-                    $this->render("image/vFormNewUpdate.html.twig", array(
-                        "title" => "Modification d'une image",
-                        "image" => $imageForm->getImage(),
-                        "action" => "update&o=image&id=" . $imageForm->getImage()->getId(),
-                        "imageForm" => $imageForm
-                    ));
                 }
             }
         }
